Pre-0.3.0 cleanup
