+++ client/.env.template
@@ -1,10 +1,11 @@
+# Default OAuth2 issuer
 AUTH0_CLIENT_ID="BnJiVrOXavZ1mbvsiwvBcZ96dTFH9k4L"
 AUTH0_ISSUER_URL="https://dev-k6zdm2f3z3kst6r7.us.auth0.com/"
 
 # For local development SpacetimeDB instance
 #DATABASE_HOST=
 
-# For maincloud SpacetimeDB instance
+# For the Official SpacetimeDB instance
 #DATABASE_HOST=https://maincloud.spacetimedb.com
 
 # Window Configuration
