+++ client/src/oidc_auth_helper.rs
@@ -1,33 +1,24 @@
-use openidconnect::{
-    AuthorizationCode,
-    ClientId,
-    CsrfToken,
-    IssuerUrl,
-    Nonce,
-    OAuth2TokenResponse,
-    PkceCodeChallenge,
-    RedirectUrl,
-    Scope,
-    TokenResponse,
-};
-use openidconnect::core::{ CoreAuthenticationFlow, CoreClient, CoreProviderMetadata };
+use openidconnect::core::{CoreAuthenticationFlow, CoreClient, CoreProviderMetadata};
 use openidconnect::reqwest;
-use url::Url;
+use openidconnect::{
+    AuthorizationCode, ClientId, CsrfToken, IssuerUrl, Nonce, OAuth2TokenResponse,
+    PkceCodeChallenge, RedirectUrl, Scope, TokenResponse,
+};
 use std::env;
+use url::Url;
 
+use std::io::{BufRead, BufReader, Write};
 use std::net::TcpListener;
-use std::io::{ BufRead, BufReader, Write };
 
 pub fn get_client_token() -> Result<String, String> {
     let auth0_client_id = ClientId::new(
-        env::var("AUTH0_CLIENT_ID").expect("Missing the AUTH0_CLIENT_ID environment variable.")
+        env::var("AUTH0_CLIENT_ID").expect("Missing the AUTH0_CLIENT_ID environment variable."),
     );
-    let issuer_url = IssuerUrl::new(
-        env::var("AUTH0_ISSUER_URL").expect("Missing AUTH0_ISSUER_URL!")
-    ).expect("Invalid issuer URL");
+    let issuer_url =
+        IssuerUrl::new(env::var("AUTH0_ISSUER_URL").expect("Missing AUTH0_ISSUER_URL!"))
+            .expect("Invalid issuer URL");
 
-    let http_client = reqwest::blocking::ClientBuilder
-        ::new()
+    let http_client = reqwest::blocking::ClientBuilder::new()
         // Following redirects opens the client up to SSRF vulnerabilities.
         .redirect(reqwest::redirect::Policy::none())
         .build()
@@ -50,7 +41,7 @@ pub fn get_client_token() -> Result<String, String> {
         .authorize_url(
             CoreAuthenticationFlow::AuthorizationCode,
             CsrfToken::new_random,
-            Nonce::new_random
+            Nonce::new_random,
         )
         // Set the desired scopes.
         .add_scope(Scope::new("read".to_string()))
@@ -91,8 +82,12 @@ pub fn get_client_token() -> Result<String, String> {
             .map(|(_, state)| CsrfToken::new(state.into_owned()))
             .unwrap();
 
-        let message =
-            "<h1>Login completed!</h1> <p>Return to <em>Solarance:Beginnings</em> and complain about this travesty of a landing page!</p>";
+        let message = {
+            let mut tmp = "<h1>Login Completed!</h1>".to_string();
+            tmp += "<p>Return to <em>Solarance:Beginnings</em> and complain about this travesty of a landing page!</p>";
+            tmp = format!("<center>{}</center>", tmp);
+            format!("<html><body>{}</body></html>", tmp)
+        };
         let response = format!(
             "HTTP/1.1 200 OK\r\ncontent-length: {}\r\n\r\n{}",
             message.len(),
@@ -118,8 +113,14 @@ pub fn get_client_token() -> Result<String, String> {
         .request(&http_client)
         .expect("Failed to contact token endpoint");
 
-    println!("Auth0 returned access token:\n{}\n", token_response.access_token().secret());
+    println!(
+        "Auth0 returned access token:\n{}\n",
+        token_response.access_token().secret()
+    );
     println!("Auth0 returned scopes: {:?}", token_response.scopes());
 
-    Ok(token_response.id_token().expect("ID token is missing or malformed.").to_string())
+    Ok(token_response
+        .id_token()
+        .expect("ID token is missing or malformed.")
+        .to_string())
 }
